refactor: 공통 components 경로 수정
- ui를 이용한 공통 components는 ui 상위 폴더에 위치
