+++ src/components/Content/index.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from 'react';
 
-import { Container, ContainerProps } from './Container';
+import { Container, ContainerProps } from '../ui/Layout/Container';
 import { contentStyle } from './styles';
 
 interface ContentProps extends ContainerProps {
