+++ src/components/ui/Layout/styles.ts
@@ -25,16 +25,6 @@ export const containerStyle = (
     gap,
   });
 
-export const contentStyle = (backgroundColor: string, height: string) =>
-  css({
-    padding: `0 1rem`,
-    backgroundColor,
-    display: 'flex',
-    justifyContent: 'center',
-    width: '100vw',
-    height,
-  });
-
 export const gridStyle = (
   gap: number,
   columns: Column,
